# Conflicts:
#	src/main/kotlin/com/taskforce/superinvention/app/domain/state/ClubState.kt
#	src/main/kotlin/com/taskforce/superinvention/app/domain/state/State.kt
#	src/main/kotlin/com/taskforce/superinvention/app/domain/state/UserState.kt

+++ src/main/kotlin/com/taskforce/superinvention/app/domain/state/UserState.kt
@@ -11,6 +11,6 @@ class UserState(
         var user: User,
         @ManyToOne
         var state: State,
-        var priority: Int
+        var priority: Long
 ) : BaseEntity() {
 }
